+++ test/unit/sql/where.test.js
@@ -345,5 +345,15 @@ suite(Support.getTestDialectTeaser('SQL'), function() {
         });
       });
     }
+
+    if (current.dialect.supports.JSON) {
+      suite('JSON', function () {
+        test('sequelize.json("profile->>\'id\', sequelize.cast(2, \'text\')")', function () {
+          expectsql(sql.whereItemQuery(undefined, this.sequelize.json("profile->>'id'", this.sequelize.cast('12346-78912', 'text'))), {
+            postgres: "profile->>'id' = CAST('12346-78912' AS TEXT)"
+          });
+        });
+      });
+    }
   });
 });